+++ lib/collection.js
@@ -6,7 +6,12 @@ const checkCollectionName = require('./utils').checkCollectionName;
 const ObjectID = require('mongodb-core').BSON.ObjectID;
 const MongoError = require('mongodb-core').MongoError;
 const toError = require('./utils').toError;
+const handleCallback = require('./utils').handleCallback;
 const normalizeHintField = require('./utils').normalizeHintField;
+const decorateCommand = require('./utils').decorateCommand;
+const decorateWithCollation = require('./utils').decorateWithCollation;
+const decorateWithReadConcern = require('./utils').decorateWithReadConcern;
+const formattedOrderClause = require('./utils').formattedOrderClause;
 const ReadPreference = require('mongodb-core').ReadPreference;
 const unordered = require('./bulk/unordered');
 const ordered = require('./bulk/ordered');
@@ -15,6 +20,7 @@ const executeOperation = require('./utils').executeOperation;
 const resolveReadPreference = require('./utils').resolveReadPreference;
 
 // Operations
+const aggregate = require('./operations/aggregate').aggregate;
 const checkForAtomicOperators = require('./operations/collection_ops').checkForAtomicOperators;
 const count = require('./operations/collection_ops').count;
 const ensureIndex = require('./operations/collection_ops').ensureIndex;
@@ -26,7 +32,6 @@ const removeDocuments = require('./operations/collection_ops').removeDocuments;
 const save = require('./operations/collection_ops').save;
 const updateDocuments = require('./operations/collection_ops').updateDocuments;
 
-const AggregateOperation = require('./operations/aggregate_operation');
 const BulkWriteOperation = require('./operations/bulk_write');
 const CountDocumentsOperation = require('./operations/count_documents');
 const CreateIndexOperation = require('./operations/create_index');
@@ -37,7 +42,6 @@ const DistinctOperation = require('./operations/distinct');
 const DropIndexOperation = require('./operations/drop_index');
 const DropIndexesOperation = require('./operations/drop_indexes');
 const EstimatedDocumentCountOperation = require('./operations/estimated_document_count');
-const FindOperation = require('./operations/find');
 const FindOneOperation = require('./operations/find_one');
 const FindOneAndDeleteOperation = require('./operations/find_one_and_delete');
 const FindOneAndReplaceOperation = require('./operations/find_one_and_replace');
@@ -85,6 +89,8 @@ const executeOperationV2 = require('./operations/execute_operation_v2');
  *   });
  * });
  */
+
+const mergeKeys = ['ignoreUndefined'];
 
 /**
  * Create a new Collection instance (INTERNAL TYPE, do not instantiate directly)
@@ -298,10 +304,135 @@ Collection.prototype.find = deprecateOptions(
 
     // Ensure selector is not null
     selector = selector == null ? {} : selector;
+    // Validate correctness off the selector
+    const object = selector;
+    if (Buffer.isBuffer(object)) {
+      const object_size = object[0] | (object[1] << 8) | (object[2] << 16) | (object[3] << 24);
+      if (object_size !== object.length) {
+        const error = new Error(
+          'query selector raw message size does not match message header size [' +
+            object.length +
+            '] != [' +
+            object_size +
+            ']'
+        );
+        error.name = 'MongoError';
+        throw error;
+      }
+    }
 
-    const findOperation = new FindOperation(this, selector, options);
+    // Check special case where we are using an objectId
+    if (selector != null && selector._bsontype === 'ObjectID') {
+      selector = { _id: selector };
+    }
 
-    return findOperation.execute(callback);
+    if (!options) options = {};
+
+    let projection = options.projection || options.fields;
+
+    if (projection && !Buffer.isBuffer(projection) && Array.isArray(projection)) {
+      projection = projection.length
+        ? projection.reduce((result, field) => {
+            result[field] = 1;
+            return result;
+          }, {})
+        : { _id: 1 };
+    }
+
+    // Make a shallow copy of options
+    let newOptions = Object.assign({}, options);
+
+    // Make a shallow copy of the collection options
+    for (let key in this.s.options) {
+      if (mergeKeys.indexOf(key) !== -1) {
+        newOptions[key] = this.s.options[key];
+      }
+    }
+
+    // Unpack options
+    newOptions.skip = options.skip ? options.skip : 0;
+    newOptions.limit = options.limit ? options.limit : 0;
+    newOptions.raw = typeof options.raw === 'boolean' ? options.raw : this.s.raw;
+    newOptions.hint =
+      options.hint != null ? normalizeHintField(options.hint) : this.s.collectionHint;
+    newOptions.timeout = typeof options.timeout === 'undefined' ? undefined : options.timeout;
+    // // If we have overridden slaveOk otherwise use the default db setting
+    newOptions.slaveOk = options.slaveOk != null ? options.slaveOk : this.s.db.slaveOk;
+
+    // Add read preference if needed
+    newOptions.readPreference = resolveReadPreference(newOptions, {
+      db: this.s.db,
+      collection: this
+    });
+
+    // Set slave ok to true if read preference different from primary
+    if (
+      newOptions.readPreference != null &&
+      (newOptions.readPreference !== 'primary' || newOptions.readPreference.mode !== 'primary')
+    ) {
+      newOptions.slaveOk = true;
+    }
+
+    // Ensure the query is an object
+    if (selector != null && typeof selector !== 'object') {
+      throw MongoError.create({ message: 'query selector must be an object', driver: true });
+    }
+
+    // Build the find command
+    const findCommand = {
+      find: this.s.namespace,
+      limit: newOptions.limit,
+      skip: newOptions.skip,
+      query: selector
+    };
+
+    // Ensure we use the right await data option
+    if (typeof newOptions.awaitdata === 'boolean') {
+      newOptions.awaitData = newOptions.awaitdata;
+    }
+
+    // Translate to new command option noCursorTimeout
+    if (typeof newOptions.timeout === 'boolean') newOptions.noCursorTimeout = newOptions.timeout;
+
+    decorateCommand(findCommand, newOptions, ['session', 'collation']);
+
+    if (projection) findCommand.fields = projection;
+
+    // Add db object to the new options
+    newOptions.db = this.s.db;
+
+    // Add the promise library
+    newOptions.promiseLibrary = this.s.promiseLibrary;
+
+    // Set raw if available at collection level
+    if (newOptions.raw == null && typeof this.s.raw === 'boolean') newOptions.raw = this.s.raw;
+    // Set promoteLongs if available at collection level
+    if (newOptions.promoteLongs == null && typeof this.s.promoteLongs === 'boolean')
+      newOptions.promoteLongs = this.s.promoteLongs;
+    if (newOptions.promoteValues == null && typeof this.s.promoteValues === 'boolean')
+      newOptions.promoteValues = this.s.promoteValues;
+    if (newOptions.promoteBuffers == null && typeof this.s.promoteBuffers === 'boolean')
+      newOptions.promoteBuffers = this.s.promoteBuffers;
+
+    // Sort options
+    if (findCommand.sort) {
+      findCommand.sort = formattedOrderClause(findCommand.sort);
+    }
+
+    // Set the readConcern
+    decorateWithReadConcern(findCommand, this, options);
+
+    // Decorate find command with collation options
+    try {
+      decorateWithCollation(findCommand, this, options);
+    } catch (err) {
+      if (typeof callback === 'function') return callback(err, null);
+      throw err;
+    }
+
+    const cursor = this.s.topology.cursor(this.s.namespace, findCommand, newOptions);
+
+    return typeof callback === 'function' ? handleCallback(callback, null, cursor) : cursor;
   }
 );
 
@@ -1569,9 +1700,7 @@ Collection.prototype.aggregate = function(pipeline, options, callback) {
     pipeline = args;
   }
 
-  const aggregateOperation = new AggregateOperation(this.s.db, this, pipeline, options);
-
-  return aggregateOperation.execute(callback);
+  return aggregate(this.s.db, this, pipeline, options, callback);
 };
 
 /**
