+++ lib/operations/bulk_write.js
@@ -2,6 +2,7 @@
 
 const applyRetryableWrites = require('../utils').applyRetryableWrites;
 const applyWriteConcern = require('../utils').applyWriteConcern;
+const defineAspects = require('./common_functions').defineAspects;
 const MongoError = require('mongodb-core').MongoError;
 const OperationBase = require('./operation').OperationBase;
 
@@ -14,7 +15,7 @@ class BulkWriteOperation extends OperationBase {
   }
 
   static get aspects() {
-    return new Set([]);
+    return this.aspects;
   }
 
   execute(callback) {
@@ -104,5 +105,7 @@ class BulkWriteOperation extends OperationBase {
     });
   }
 }
+
+defineAspects(BulkWriteOperation, []);
 
 module.exports = BulkWriteOperation;
