+++ lib/operations/common_functions.js
@@ -9,6 +9,15 @@ const handleCallback = require('../utils').handleCallback;
 const ReadPreference = require('mongodb-core').ReadPreference;
 const toError = require('../utils').toError;
 
+function defineAspects(operation, aspects) {
+  aspects = new Set(aspects);
+  Object.defineProperty(operation, 'aspects', {
+    value: aspects,
+    writable: false
+  });
+  return aspects;
+}
+
 function deleteCallback(err, r, callback) {
   if (callback == null) return;
   if (err && callback) return callback(err);
@@ -241,6 +250,7 @@ function updateCallback(err, r, callback) {
 }
 
 module.exports = {
+  defineAspects,
   deleteCallback,
   findAndModify,
   prepareDocs,
