+++ lib/operations/delete_many.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const OperationBase = require('./operation').OperationBase;
+const defineAspects = require('./common_functions').defineAspects;
 const deleteCallback = require('./common_functions').deleteCallback;
 const removeDocuments = require('./common_functions').removeDocuments;
 
@@ -13,7 +14,7 @@ class DeleteManyOperation extends OperationBase {
   }
 
   static get aspects() {
-    return new Set([]);
+    return this.aspects;
   }
 
   execute(callback) {
@@ -25,5 +26,7 @@ class DeleteManyOperation extends OperationBase {
     removeDocuments(coll, filter, options, (err, r) => deleteCallback(err, r, callback));
   }
 }
+
+defineAspects(DeleteManyOperation, []);
 
 module.exports = DeleteManyOperation;
