+++ lib/operations/delete_one.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const OperationBase = require('./operation').OperationBase;
+const defineAspects = require('./common_functions').defineAspects;
 const deleteCallback = require('./common_functions').deleteCallback;
 const removeDocuments = require('./common_functions').removeDocuments;
 
@@ -25,5 +26,7 @@ class DeleteOneOperation extends OperationBase {
     removeDocuments(coll, filter, options, (err, r) => deleteCallback(err, r, callback));
   }
 }
+
+defineAspects(DeleteOneOperation, []);
 
 module.exports = DeleteOneOperation;
