+++ lib/operations/find_one.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const defineAspects = require('./common_functions').defineAspects;
 const handleCallback = require('../utils').handleCallback;
 const OperationBase = require('./operation').OperationBase;
 const toError = require('../utils').toError;
@@ -13,7 +14,7 @@ class FindOneOperation extends OperationBase {
   }
 
   static get aspects() {
-    return new Set([]);
+    return this.aspects;
   }
 
   execute(callback) {
@@ -33,5 +34,7 @@ class FindOneOperation extends OperationBase {
     });
   }
 }
+
+defineAspects(FindOneOperation, []);
 
 module.exports = FindOneOperation;
