+++ lib/operations/operation.js
@@ -5,8 +5,9 @@ const Aspect = {
 };
 
 class OperationBase {
-  constructor(options) {
+  constructor(options, aspects) {
     this.options = options || {};
+    this.aspects = aspects || new Set([]);
   }
 
   hasAspect(aspect) {
