+++ lib/operations/options_operation.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const OperationBase = require('./operation').OperationBase;
+const defineAspects = require('./common_functions').defineAspects;
 const handleCallback = require('../utils').handleCallback;
 const MongoError = require('mongodb-core').MongoError;
 
@@ -12,7 +13,7 @@ class OptionsOperation extends OperationBase {
   }
 
   static get aspects() {
-    return new Set([]);
+    return this.aspects;
   }
 
   execute(callback) {
@@ -32,5 +33,7 @@ class OptionsOperation extends OperationBase {
     });
   }
 }
+
+defineAspects(OptionsOperation, []);
 
 module.exports = OptionsOperation;
