+++ lib/operations/replace_one.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const OperationBase = require('./operation').OperationBase;
+const defineAspects = require('./common_functions').defineAspects;
 const updateDocuments = require('./common_functions').updateDocuments;
 
 class ReplaceOneOperation extends OperationBase {
@@ -13,7 +14,7 @@ class ReplaceOneOperation extends OperationBase {
   }
 
   static get aspects() {
-    return new Set([]);
+    return this.aspects;
   }
 
   execute(callback) {
@@ -47,5 +48,7 @@ function replaceCallback(err, r, doc, callback) {
   r.ops = [doc];
   if (callback) callback(null, r);
 }
+
+defineAspects(ReplaceOneOperation, []);
 
 module.exports = ReplaceOneOperation;
