+++ lib/operations/update_many.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const OperationBase = require('./operation').OperationBase;
+const defineAspects = require('./common_functions').defineAspects;
 const updateCallback = require('./common_functions').updateCallback;
 const updateDocuments = require('./common_functions').updateDocuments;
 
@@ -14,7 +15,7 @@ class UpdateManyOperation extends OperationBase {
   }
 
   static get aspects() {
-    return new Set([]);
+    return this.aspects;
   }
 
   execute(callback) {
@@ -29,5 +30,7 @@ class UpdateManyOperation extends OperationBase {
     updateDocuments(coll, filter, update, options, (err, r) => updateCallback(err, r, callback));
   }
 }
+
+defineAspects(UpdateManyOperation, []);
 
 module.exports = UpdateManyOperation;
