+++ lib/operations/update_one.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const OperationBase = require('./operation').OperationBase;
+const defineAspects = require('./common_functions').defineAspects;
 const updateDocuments = require('./common_functions').updateDocuments;
 
 class UpdateOneOperation extends OperationBase {
@@ -13,7 +14,7 @@ class UpdateOneOperation extends OperationBase {
   }
 
   static get aspects() {
-    return new Set([]);
+    return this.aspects;
   }
 
   execute(callback) {
@@ -44,5 +45,7 @@ function updateCallback(err, r, callback) {
     Array.isArray(r.result.upserted) && r.result.upserted.length > 0 ? 0 : r.result.n;
   callback(null, r);
 }
+
+defineAspects(UpdateOneOperation, []);
 
 module.exports = UpdateOneOperation;
